requirements.txt
Django==4.1.7
djangorestframework==3.14.0
Pillow==9.4.0
django-jazzmin==2.6.0
python-dotenv==1.0.0
django-modeltranslation==0.18.9